Dynamically set TERM=tmux-direct
I'd previously been using screen-256color as my default terminal in
tmux, which basically worked.  But tmux-256color seems to be available
everywhere I care about, so this makes that the default instead.

Then, in .zshrc, I now check whether TERM = tmux-256color to infer
whether we're running in tmux, and if so dynamically check for the
availablity of the tmux-direct terminfo.  If it's available I use that
instead.