[vimrc] Configure vim to work in true color
Some settings need to be changed from their defaults so Vim will operate
correctly in -direct terminals, especially in tmux-direct.

This configuration currently hard-codes Vim to assume we're running with
a dark background color.  Some further work may be needed for programs
running in tmux to automatically detect whether they have a light or a
dark background color (and this may apply to Emacs too).